Move Signup form to tcomb
